+++ components/auto-complete/auto-complete.vue
@@ -265,6 +265,7 @@ function handleInputInternal(event: string | Event) {
 
   emitEvent(props.onInput, value)
   nextTick(testOptionCanDrop)
+  select.value?.updatePopper()
 }
 
 const delay = toNumber(props.delay)