+++ components/select/select.vue
@@ -599,6 +599,7 @@ defineExpose({
   virtualList,
   restTip,
 
+  updatePopper,
   isSelected,
   getOptionFromMap,
   updateHitting,
@@ -690,6 +691,7 @@ function initHittingIndex() {
 }
 
 function setVisible(visible: boolean) {
+  console.trace('1')
   if (currentVisible.value === visible) return
 
   currentVisible.value = visible
@@ -1018,7 +1020,12 @@ function handleCompositionEnd() {
 function handleFilterKeyDown(event: KeyboardEvent) {
   if (!input.value) return
 
-  if (event.key === 'Backspace' && !input.value.value && !isNull(getLast(currentValues.value))) {
+  if (
+    props.filterPosition !== 'in-list' &&
+    event.key === 'Backspace' &&
+    !input.value.value &&
+    !isNull(getLast(currentValues.value))
+  ) {
     event.stopPropagation()
     handleTagClose(getLast(currentValues.value))
   }